fix the following warnings, reported by gcc version 4.1.2 20080704 (Red Hat 4.1.2-48) on CentOS 5.5/x86_64: {{{ gfsd.c: In function 'file_table_close': gfsd.c:739: warning: format '%lld' expects type 'long long int', but argument 9 has type 'gfarm_ino_t' gfsd.c:739: warning: format '%lld' expects type 'long long int', but argument 10 has type 'gfarm_uint64_t' gfsd.c:739: warning: format '%lld' expects type 'long long int', but argument 11 has type 'gfarm_uint64_t' gfsd.c:739: warning: format '%lld' expects type 'long long int', but argument 13 has type 'gfarm_off_t' gfsd.c:739: warning: format '%lld' expects type 'long long int', but argument 16 has type 'gfarm_off_t' }}}
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@5445 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ server/gfsd/gfsd.c
@@ -752,9 +752,11 @@ file_table_close(gfarm_int32_t net_fd)
 		    "inum %lld gen %lld (%lld) "
 		    "write %d size %lld time %g "
 		    "read %d size %lld time %g",
-		    time_buf, total_time, fe->ino, fe->gen, fe->new_gen,
-		    fe->nwrite, fe->write_size, fe->write_time,
-		    fe->nread, fe->read_size, fe->read_time));
+		    time_buf, total_time, (unsigned long long)fe->ino,
+		    (unsigned long long)fe->gen,
+		    (unsigned long long)fe->new_gen,
+		    fe->nwrite, (long long)fe->write_size, fe->write_time,
+		    fe->nread, (long long)fe->read_size, fe->read_time));
 
 	return (e);
 }
